server/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
pymongo[srv]==4.5.0
motor==3.2.0
pydantic==2.0.3
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.25.0
